# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Error {
  error: string;
}

export interface City {
  id: string;
  name: string;
  governorate: string;
  tagline: string;
  questionCount: number;
}

export interface QuizOption {
  id: string;
  label: string;
}

export type QuizQuestionKind = typeof QuizQuestionKind[keyof typeof QuizQuestionKind];


export const QuizQuestionKind = {
  street: 'street',
  landmark: 'landmark',
} as const;

export interface QuizQuestion {
  id: string;
  prompt: string;
  kind: QuizQuestionKind;
  /**
     * @minimum -90
     * @maximum 90
     */
  targetLat: number;
  /**
     * @minimum -180
     * @maximum 180
     */
  targetLng: number;
  options: QuizOption[];
}

export interface MapBounds {
  /**
     * @minimum -90
     * @maximum 90
     */
  north: number;
  /**
     * @minimum -90
     * @maximum 90
     */
  south: number;
  /**
     * @minimum -180
     * @maximum 180
     */
  east: number;
  /**
     * @minimum -180
     * @maximum 180
     */
  west: number;
}

export interface CityQuiz {
  city: City;
  questions: QuizQuestion[];
  mapBounds: MapBounds;
}

export interface QuizAnswer {
  questionId: string;
  optionId: string;
}

export interface QuizSubmission {
  cityId: string;
  answers: QuizAnswer[];
}

export interface QuizAnswerResult {
  questionId: string;
  selectedOptionId: string;
  correctOptionId: string;
  correctLabel: string;
  isCorrect: boolean;
}

export interface QuizResult {
  cityName: string;
  correct: number;
  total: number;
  scorePercent: number;
  grade: string;
  results: QuizAnswerResult[];
}