# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: mapQuiz
    description: City map quiz operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /cities:
    get:
      operationId: listCities
      tags: [mapQuiz]
      summary: List available cities
      responses:
        "200":
          description: Available Egyptian cities
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/City"
  /cities/{cityId}/quiz:
    get:
      operationId: getCityQuiz
      tags: [mapQuiz]
      summary: Get a city's quiz questions
      parameters:
        - name: cityId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Quiz questions for the selected city
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CityQuiz"
        "404":
          description: City not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /quiz/score:
    post:
      operationId: submitQuiz
      tags: [mapQuiz]
      summary: Grade a completed quiz
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuizSubmission"
      responses:
        "200":
          description: Graded quiz result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuizResult"
        "400":
          description: Invalid quiz answers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: City not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    City:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        governorate:
          type: string
        tagline:
          type: string
        questionCount:
          type: integer
      required: [id, name, governorate, tagline, questionCount]
    QuizOption:
      type: object
      properties:
        id:
          type: string
        label:
          type: string
      required: [id, label]
    QuizQuestion:
      type: object
      properties:
        id:
          type: string
        prompt:
          type: string
        kind:
          type: string
          enum: [street, landmark]
        targetLat:
          type: number
          minimum: -90
          maximum: 90
        targetLng:
          type: number
          minimum: -180
          maximum: 180
        options:
          type: array
          items:
            $ref: "#/components/schemas/QuizOption"
      required: [id, prompt, kind, targetLat, targetLng, options]
    MapBounds:
      type: object
      properties:
        north:
          type: number
          minimum: -90
          maximum: 90
        south:
          type: number
          minimum: -90
          maximum: 90
        east:
          type: number
          minimum: -180
          maximum: 180
        west:
          type: number
          minimum: -180
          maximum: 180
      required: [north, south, east, west]
    CityQuiz:
      type: object
      properties:
        city:
          $ref: "#/components/schemas/City"
        questions:
          type: array
          items:
            $ref: "#/components/schemas/QuizQuestion"
        mapBounds:
          $ref: "#/components/schemas/MapBounds"
      required: [city, questions, mapBounds]
    QuizAnswer:
      type: object
      properties:
        questionId:
          type: string
        optionId:
          type: string
      required: [questionId, optionId]
    QuizSubmission:
      type: object
      properties:
        cityId:
          type: string
        answers:
          type: array
          items:
            $ref: "#/components/schemas/QuizAnswer"
      required: [cityId, answers]
    QuizAnswerResult:
      type: object
      properties:
        questionId:
          type: string
        selectedOptionId:
          type: string
        correctOptionId:
          type: string
        correctLabel:
          type: string
        isCorrect:
          type: boolean
      required: [questionId, selectedOptionId, correctOptionId, correctLabel, isCorrect]
    QuizResult:
      type: object
      properties:
        cityName:
          type: string
        correct:
          type: integer
        total:
          type: integer
        scorePercent:
          type: integer
        grade:
          type: string
        results:
          type: array
          items:
            $ref: "#/components/schemas/QuizAnswerResult"
      required: [cityName, correct, total, scorePercent, grade, results]