# lib/api-zod/src/generated/types/mapBounds.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MapBounds {
  /**
     * @minimum -90
     * @maximum 90
     */
  north: number;
  /**
     * @minimum -90
     * @maximum 90
     */
  south: number;
  /**
     * @minimum -180
     * @maximum 180
     */
  east: number;
  /**
     * @minimum -180
     * @maximum 180
     */
  west: number;
}