# lib/api-zod/src/generated/types/quizQuestion.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { QuizOption } from './quizOption';
import type { QuizQuestionKind } from './quizQuestionKind';

export interface QuizQuestion {
  id: string;
  prompt: string;
  kind: QuizQuestionKind;
  /**
     * @minimum -90
     * @maximum 90
     */
  targetLat: number;
  /**
     * @minimum -180
     * @maximum 180
     */
  targetLng: number;
  options: QuizOption[];
}